pyproject.toml
[tool.poetry]
name = "ghostpost_backend"
version = "0.1.0"
description = ""
authors = ["AlvaradoM <mauriciomj11@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
django-cors-headers = "^3.4.0"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
